feat(service): Add metricas

+++ src/main/java/com/example/SalesHub/dto/response/entity/EstoqueResponse.java
@@ -1,9 +1,10 @@
 package com.example.SalesHub.dto.response.entity;
 
 import lombok.Builder;
+import java.math.BigDecimal;
 
 @Builder
-public record EstoqueResponse( Long id,
-                                                              Long produtoId,
-                                                              Long quantidadeInicial,
-                                                              Long quantidadeAtual) {}
+public record EstoqueResponse(Long id,
+                              Long produtoId,
+                              BigDecimal quantidadeInicial,
+                              BigDecimal quantidadeAtual) {}
